+++ src/common/consts/routes.ts
@@ -6,8 +6,6 @@ export const ROUTES = {
   forgotPassword: '/forgot-password',
   learn: '/learn',
   profile: '/profile',
-  // recoverPassword: '/recover-password',
-  rest: '/*',
   signIn: '/sign-in',
   signUp: '/sign-up',
 } as const
