+++ src/components/layouts/header/header.tsx
@@ -1,11 +1,50 @@
 import { memo } from 'react'
+import { Link } from 'react-router-dom'
 
-import { useNoAuthRoutes } from '@/common/hooks/no-auth-routes'
-import { PrivateHeader } from '@/components/layouts/header/privatHeader'
-import { PublicHeader } from '@/components/layouts/header/publicHeader'
+import logo from '@/assets/images/LogoITI.svg'
+import { ROUTES } from '@/common/consts/routes'
+import { MenuProfile } from '@/components/menus'
+import { Avatar, Button, Typography } from '@/components/ui'
+import { GetUser } from '@/services/auth/auth.types'
 
-export const Header = memo(() => {
-  const isNoAuthRoute = useNoAuthRoutes()
+import s from '@/components/layouts/header/header.module.scss'
 
-  return isNoAuthRoute ? <PublicHeader /> : <PrivateHeader />
+type Props = {
+  data: GetUser | null
+  logout: () => void
+}
+
+export const Header = memo(({ data, logout }: Props) => {
+  const triggerHeader = (
+    <div className={s.nameContainer}>
+      <Typography.Subtitle1 as={Link} className={s.name} to={ROUTES.profile}>
+        {data?.name}
+      </Typography.Subtitle1>
+      <Avatar name={data?.name} size={'s'} url={data?.avatar} />
+    </div>
+  )
+
+  return (
+    <header className={s.root}>
+      <div className={s.content}>
+        <Link className={s.banner} to={ROUTES.base}>
+          <img alt={'Logo'} height={36} src={logo} width={157} />
+        </Link>
+
+        {data ? (
+          <MenuProfile
+            avatarUrl={data.avatar}
+            email={data.email}
+            onLogout={logout}
+            triggerMenu={triggerHeader}
+            userName={data.name}
+          />
+        ) : (
+          <Button as={Link} to={ROUTES.signIn} variant={'secondary'}>
+            Sign In
+          </Button>
+        )}
+      </div>
+    </header>
+  )
 })
