--- src/components/layouts/header/privatHeader.tsx
@@ -1,57 +1,0 @@
-import { Link, Navigate } from 'react-router-dom'
-
-import logo from '@/assets/images/LogoITI.svg'
-import { ROUTES } from '@/common/consts/routes'
-import { useSuccessNotification } from '@/common/hooks/use-success-notification'
-import { MenuProfile } from '@/components/menus'
-import { Avatar, Typography } from '@/components/ui'
-import { useGetMeQuery, useLogoutMutation } from '@/services/auth/auth.service'
-
-import s from './header.module.scss'
-
-export const PrivateHeader = () => {
-  const { data } = useGetMeQuery()
-  const [logout, { isSuccess: success }] = useLogoutMutation()
-
-  const handleLogout = async () => {
-    await logout()
-  }
-
-  const user = {
-    avatar: data?.avatar || '',
-    email: data?.email || '',
-    name: data?.name || '',
-  }
-
-  useSuccessNotification(success, 'You are successfully logged out')
-
-  if (success) {
-    return <Navigate replace to={ROUTES.signIn} />
-  }
-
-  const triggerHeader = (
-    <div className={s.nameContainer}>
-      <Typography.Subtitle1 as={Link} className={s.name} to={ROUTES.profile}>
-        {user.name}
-      </Typography.Subtitle1>
-      <Avatar name={user.name} size={'s'} url={user.avatar} />
-    </div>
-  )
-
-  return (
-    <header className={s.root}>
-      <div className={s.content}>
-        <Link className={s.banner} to={ROUTES.base}>
-          <img alt={'Logo'} height={36} src={logo} width={157} />
-        </Link>
-        <MenuProfile
-          avatarUrl={user.avatar}
-          email={user.email}
-          onLogout={handleLogout}
-          triggerMenu={triggerHeader}
-          userName={user.name}
-        />
-      </div>
-    </header>
-  )
-}
