+++ src/components/layouts/layout/layout.tsx
@@ -1,9 +1,11 @@
 import { ComponentPropsWithoutRef, ElementRef, forwardRef } from 'react'
 import { useSelector } from 'react-redux'
+import { Outlet } from 'react-router-dom'
 
 import { Header } from '@/components/layouts/header/header'
 import { Loader } from '@/components/ui'
 import { selectAppIsLoading } from '@/services/app/app.selectors'
+import { useGetMeQuery, useLogoutMutation } from '@/services/auth/auth.service'
 import clsx from 'clsx'
 
 import s from './layout.module.scss'
@@ -12,6 +14,17 @@ type Props = ComponentPropsWithoutRef<'div'>
 
 export const Layout = forwardRef<ElementRef<'div'>, Props>(
   ({ children, className, ...rest }, ref) => {
+    const { data } = useGetMeQuery()
+
+    const [logout] = useLogoutMutation()
+
+    const headerData =
+      data && !('success' in data)
+        ? {
+            ...data,
+          }
+        : null
+
     const isLoading = useSelector(selectAppIsLoading)
 
     const classNames = {
@@ -22,8 +35,10 @@ export const Layout = forwardRef<ElementRef<'div'>, Props>(
     return (
       <div ref={ref} {...rest} className={classNames.root}>
         {isLoading && <Loader />}
-        <Header />
-        <main className={classNames.main}>{children}</main>
+        <Header data={headerData} logout={logout} />
+        <main className={classNames.main}>
+          <Outlet />
+        </main>
       </div>
     )
   }
